# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Documentation",
    "description": "\n",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWTAuth": {
        "type": "http",
        "description": "JWT Authentication Scheme",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "AnyObject": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "__": {
            "type": "string"
          }
        },
        "required": [
          "__"
        ],
        "additionalProperties": false,
        "id": "AnyObject"
      },
      "PaginationMetaData": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ],
        "additionalProperties": false,
        "id": "PaginationMetaData"
      },
      "BusData": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
          },
          "licensePlate": {
            "type": "string"
          },
          "capacity": {
            "type": "number"
          },
          "metadata": {
            "$ref": "#/components/schemas/AnyObject"
          }
        },
        "required": [
          "id",
          "licensePlate",
          "capacity",
          "metadata"
        ],
        "additionalProperties": false,
        "id": "BusData"
      },
...
    }
  },
  "paths": {
...
    "/api/buses/": {
      "get": {
        "tags": [
          "BusesController"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "description": "Search term to filter results",
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "default": 1,
              "description": "Page number, minimum is 1",
              "type": "number",
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "default": 10,
              "description": "Number of items per page, minimum is 1 and maximum is 100",
              "type": "number",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 200,
                      "maximum": 300
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "data": {
                      "$schema": "https://json-schema.org/draft/2020-12/schema",
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/BusData"
                          }
                        },
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMetaData"
                        }
                      },
                      "required": [
                        "data",
                        "meta"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "title": "ApiErrorResponse",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "name": {
                      "type": "string",
                      "description": "Error class name"
                    }
                  },
                  "required": []
                }
              }
            }
          }
        },
        "security": [
          {
            "JWTAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "BusesController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "licensePlate": {
                    "type": "string"
                  },
                  "capacity": {
                    "type": "number"
                  },
                  "metadata": {
                    "$ref": "#/components/schemas/AnyObject"
                  }
                },
                "required": [
                  "licensePlate",
                  "capacity",
                  "metadata"
                ],
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 200,
                      "maximum": 300
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "data": {
                      "$schema": "https://json-schema.org/draft/2020-12/schema",
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                        },
                        "licensePlate": {
                          "type": "string"
                        },
                        "capacity": {
                          "type": "number"
                        },
                        "metadata": {
                          "$ref": "#/components/schemas/AnyObject"
                        }
                      },
                      "required": [
                        "id",
                        "licensePlate",
                        "capacity",
                        "metadata"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "title": "ApiErrorResponse",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "name": {
                      "type": "string",
                      "description": "Error class name"
                    }
                  },
                  "required": []
                }
              }
            }
          }
        },
        "security": [
          {
            "JWTAuth": []
          }
        ]
      }
    },
    "/api/buses/{id}": {
      "get": {
        "tags": [
          "BusesController"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 200,
                      "maximum": 300
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "data": {
                      "$schema": "https://json-schema.org/draft/2020-12/schema",
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
                        },
                        "licensePlate": {
                          "type": "string"
                        },
                        "capacity": {
                          "type": "number"
                        },
                        "metadata": {
                          "$ref": "#/components/schemas/AnyObject"
                        },
                        "test": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "licensePlate",
                        "capacity",
                        "metadata",
                        "test"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "title": "ApiErrorResponse",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "name": {
                      "type": "string",
                      "description": "Error class name"
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "title": "ApiErrorResponse",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "name": {
                      "type": "string",
                      "description": "Error class name"
                    }
                  },
                  "required": []
                }
              }
            }
          }
        },
        "security": [
          {
            "JWTAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "BusesController"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 200,
                      "maximum": 300
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "title": "ApiErrorResponse",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "name": {
                      "type": "string",
                      "description": "Error class name"
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "title": "ApiErrorResponse",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "name": {
                      "type": "string",
                      "description": "Error class name"
                    }
                  },
                  "required": []
                }
              }
            }
          }
        },
        "security": [
          {
            "JWTAuth": []
          }
        ]
      }
    },
...
  }
}